// apps/manager/prisma/generated/zod/inputTypeSchemas/McpServerUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { McpServerUpdateargsInputSchema } from './McpServerUpdateargsInputSchema';
import { McpServerUpdateenvVarsInputSchema } from './McpServerUpdateenvVarsInputSchema';
import { BoolFieldUpdateOperationsInputSchema } from './BoolFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { ToolUpdateManyWithoutMcpServerNestedInputSchema } from './ToolUpdateManyWithoutMcpServerNestedInputSchema';
import { UserMcpServerConfigUpdateManyWithoutMcpServerNestedInputSchema } from './UserMcpServerConfigUpdateManyWithoutMcpServerNestedInputSchema';

export const McpServerUpdateInputSchema: z.ZodType<Prisma.McpServerUpdateInput> = z.object({
  id: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  iconPath: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  command: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  args: z.union([ z.lazy(() => McpServerUpdateargsInputSchema),z.string().array() ]).optional(),
  envVars: z.union([ z.lazy(() => McpServerUpdateenvVarsInputSchema),z.string().array() ]).optional(),
  isPublic: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  tools: z.lazy(() => ToolUpdateManyWithoutMcpServerNestedInputSchema).optional(),
  mcpServerConfigs: z.lazy(() => UserMcpServerConfigUpdateManyWithoutMcpServerNestedInputSchema).optional()
}).strict();

export default McpServerUpdateInputSchema;